move events to model events .cs - just some cleanup resx for add style resx for restart resx for tooltip on restart button
