+++ src/DynamoCore/Models/DynamoModel.cs
@@ -119,132 +119,6 @@ namespace Dynamo.Models
 
         #endregion
 
-        #region events
-
-        internal delegate void FunctionNamePromptRequestHandler(object sender, FunctionNamePromptEventArgs e);
-        internal event FunctionNamePromptRequestHandler RequestsFunctionNamePrompt;
-        internal void OnRequestsFunctionNamePrompt(Object sender, FunctionNamePromptEventArgs e)
-        {
-            if (RequestsFunctionNamePrompt != null)
-                RequestsFunctionNamePrompt(this, e);
-        }
-
-        internal event Action<PresetsNamePromptEventArgs> RequestPresetsNamePrompt;
-        internal void OnRequestPresetNamePrompt(PresetsNamePromptEventArgs e)
-        {
-            if (RequestPresetsNamePrompt != null)
-                RequestPresetsNamePrompt(e);
-        }
-
-        /// <summary>
-        /// Occurs when a workspace is saved to a file.
-        /// </summary>
-        public event WorkspaceHandler WorkspaceSaved;
-        internal void OnWorkspaceSaved(WorkspaceModel workspace)
-        {
-            if (WorkspaceSaved != null)
-            {
-                WorkspaceSaved(workspace);
-            }
-        }
-
-        /// <summary>
-        /// Occurs when a workspace is about to be saved to a file.
-        /// </summary>
-        public event WorkspaceSaveHandler WorkspaceSaving;
-        internal void OnWorkspaceSaving(WorkspaceModel workspace, SaveContext saveContext)
-        {
-            if (WorkspaceSaving != null)
-            {
-                WorkspaceSaving(workspace, saveContext);
-                if (workspace is HomeWorkspaceModel hws)
-                    HandleStorageExtensionsOnWorkspaceSaving(hws, saveContext);
-            }
-        }
-
-        /// <summary>
-        /// Occurs when a workspace is scheduled to be saved to a backup file.
-        /// </summary>
-        public event Action<string, bool> RequestWorkspaceBackUpSave;
-        internal void OnRequestWorkspaceBackUpSave(string path, bool isBackUp)
-        {
-            if (RequestWorkspaceBackUpSave != null)
-            {
-                RequestWorkspaceBackUpSave(path, isBackUp);
-            }
-        }
-
-        /// <summary>
-        /// Event that is fired during the opening of the workspace.
-        ///
-        /// Use the XmlDocument object provided to conduct additional
-        /// workspace opening operations.
-        /// </summary>
-        public event Action<object> WorkspaceOpening;
-        internal void OnWorkspaceOpening(object obj)
-        {
-            var handler = WorkspaceOpening;
-            if (handler != null) handler(obj);
-        }
-
-        /// <summary>
-        /// Occurs when a workspaces is opened
-        /// </summary>
-        public event WorkspaceHandler WorkspaceOpened;
-        internal void OnWorkspaceOpened(WorkspaceModel workspace)
-        {
-            if (WorkspaceOpened != null)
-            {
-                WorkspaceOpened.Invoke(workspace);
-                if (workspace is HomeWorkspaceModel hws)
-                    HandleStorageExtensionsOnWorkspaceOpened(hws);
-            }
-        }
-
-        /// <summary>
-        /// This event is raised right before the shutdown of DynamoModel started.
-        /// When this event is raised, the shutdown is guaranteed to take place
-        /// (i.e. user has had a chance to save the work and decided to proceed
-        /// with shutting down Dynamo). Handlers of this event can still safely
-        /// access the DynamoModel, the WorkspaceModel (along with its contents),
-        /// and the DynamoScheduler.
-        /// </summary>
-        public event DynamoModelHandler ShutdownStarted;
-
-        private void OnShutdownStarted()
-        {
-            if (ShutdownStarted != null)
-                ShutdownStarted(this);
-        }
-
-        /// <summary>
-        /// This event is raised after DynamoModel has been shut down. At this
-        /// point the DynamoModel is no longer valid and access to it should be
-        /// avoided.
-        /// </summary>
-        public event DynamoModelHandler ShutdownCompleted;
-
-        private void OnShutdownCompleted()
-        {
-            if (ShutdownCompleted != null)
-                ShutdownCompleted(this);
-        }
-
-        /// <summary>
-        /// This event is raised when Dynamo is ready for user interaction.
-        /// </summary>
-        public event Action<ReadyParams> DynamoReady;
-        private bool dynamoReady;
-
-        internal static event Action<string> RequestPythonRestart;
-        internal void OnRequestPythonRestart(string pythonEngine)
-        {
-            RequestPythonRestart?.Invoke(pythonEngine);
-            //TODO now mark everything dirty.
-        }    
-
-        #endregion
-
         #region static properties
 
         /// <summary>
@@ -815,7 +689,7 @@ namespace Dynamo.Models
             var extensions = config.Extensions ?? LoadExtensions();
 
             LinterManager = new LinterManager(this.ExtensionManager);
-            
+
             // when dynamo is ready, alert the loaded extensions
             DynamoReady += (readyParams) =>
             {
@@ -1050,8 +924,8 @@ namespace Dynamo.Models
                 logger.Log(ex.Message + " : " + ex.StackTrace);
                 return;
             }
-            
-            
+
+
             if (hasMatchingExtensionData)
             {
                 workspace.UpdateExtensionData(extension.UniqueId, data);
@@ -1183,7 +1057,7 @@ namespace Dynamo.Models
                                     // Tracking node execution as generic event
                                     // it is distinguished with the legacy aggregated performance event
                                     Dynamo.Logging.Analytics.TrackEvent(
-                                        Actions.Run, 
+                                        Actions.Run,
                                         Categories.NodeOperations,
                                         node.GetOriginalName());
                                 }
@@ -1473,7 +1347,7 @@ namespace Dynamo.Models
 
         private void InitializeAnalyticsService()
         {
-           AnalyticsService.Start(this,disableADP, IsHeadless, IsTestMode);
+            AnalyticsService.Start(this, disableADP, IsHeadless, IsTestMode);
         }
 
         private IPreferences CreateOrLoadPreferences(IPreferences preferences)
@@ -2248,7 +2122,7 @@ namespace Dynamo.Models
                 IsTestMode, LinterManager, string.Empty);
 
             defaultWorkspace.RunSettings.RunType = PreferenceSettings.DefaultRunType;
-            
+
             RegisterHomeWorkspace(defaultWorkspace);
             AddWorkspace(defaultWorkspace);
             CurrentWorkspace = defaultWorkspace;
@@ -2454,7 +2328,7 @@ namespace Dynamo.Models
 
             // Create the new NodeModel's
             var newNodeModels = new List<NodeModel>();
-            using(CurrentWorkspace.BeginDelayedGraphExecution())
+            using (CurrentWorkspace.BeginDelayedGraphExecution())
             {
                 foreach (var node in nodes)
                 {
@@ -2523,17 +2397,17 @@ namespace Dynamo.Models
 
                         // If the guid is in nodeLookup, then we connect to the new pasted node. Otherwise we
                         // re-connect to the original.
-                let startNode =
-                        modelLookup.TryGetValue(c.Start.Owner.GUID, out start)
-                            ? start as NodeModel
-                            : CurrentWorkspace.Nodes.FirstOrDefault(x => x.GUID == c.Start.Owner.GUID)
+                    let startNode =
+                            modelLookup.TryGetValue(c.Start.Owner.GUID, out start)
+                                ? start as NodeModel
+                                : CurrentWorkspace.Nodes.FirstOrDefault(x => x.GUID == c.Start.Owner.GUID)
                     let endNode =
                         modelLookup.TryGetValue(c.End.Owner.GUID, out end)
                             ? end as NodeModel
                             : CurrentWorkspace.Nodes.FirstOrDefault(x => x.GUID == c.End.Owner.GUID)
 
-                // Don't make a connector if either end is null.
-                where startNode != null && endNode != null
+                    // Don't make a connector if either end is null.
+                    where startNode != null && endNode != null
                     select
                         ConnectorModel.Make(startNode, endNode, c.Start.Index, c.End.Index);
 
@@ -2593,7 +2467,7 @@ namespace Dynamo.Models
 
                 // Record models that are created as part of the command.
                 CurrentWorkspace.RecordCreatedModels(createdModels);
-            }    
+            }
         }
 
         /// <summary>
@@ -2734,7 +2608,7 @@ namespace Dynamo.Models
 
             Action savedHandler = () => OnWorkspaceSaved(workspace);
             workspace.Saved += savedHandler;
-            Action<SaveContext> savingHandler = (c) => OnWorkspaceSaving(workspace,c);
+            Action<SaveContext> savingHandler = (c) => OnWorkspaceSaving(workspace, c);
             workspace.WorkspaceSaving += savingHandler;
             workspace.MessageLogged += LogMessage;
             workspace.PropertyChanged += OnWorkspacePropertyChanged;
