+++ src/DynamoCoreWpf/Properties/Resources.en-US.resx
@@ -2390,4 +2390,13 @@ Uninstall the following packages: {0}?</value>
   <data name="PreferencesViewEnableNodeAutoCompleteTooltipText" xml:space="preserve">
     <value>Learn more about Node Autocomplete feature.</value>
   </data>
+  <data name="AddStyleButton" xml:space="preserve">
+    <value>Add Style</value>
+  </data>
+  <data name="RestartCPythonButtonText" xml:space="preserve">
+    <value>Restart CPython</value>
+  </data>
+  <data name="RestartCPythonButtonToolTip" xml:space="preserve">
+    <value>Restart the CPython engine and reload python modules.</value>
+  </data>
 </root>
